Adapt package layout for new services to project convention
