+++ src/main/scala/com/greencatsoft/angularjs/anchorScroll/AnchorScroll.scala
@@ -1,0 +1,9 @@
+package com.greencatsoft.angularjs.anchorScroll
+
+import scala.scalajs.js
+
+trait AnchorScroll extends js.Function0[Unit]
+
+object AnchorScroll {
+  val Name = "$anchorScroll"
+}
