+++ src/main/scala/com/greencatsoft/angularjs/animate/Animate.scala
@@ -1,6 +1,4 @@
-package com.greencatsoft.angularjs.services
-
-import com.greencatsoft.angularjs.InjectionTarget
+package com.greencatsoft.angularjs.animate
 
 import scala.scalajs.js
 
@@ -23,15 +21,3 @@ object Animate {
   val Name = "$animate"
 }
 
-trait AnimateAware extends InjectionTarget {
-  implicit var animate: Animate = _
-
-  override def dependencies = super.dependencies :+ Animate.Name
-
-  override def inject(args: Seq[js.Any]) {
-    super.inject(args)
-
-    var index = dependencies.indexOf(Animate.Name) ensuring (_ >= 0)
-    this.animate = args(index).asInstanceOf[Animate]
-  }
-}
