+++ src/main/scala/com/greencatsoft/angularjs/animate/AnimateAware.scala
@@ -1,0 +1,17 @@
+package com.greencatsoft.angularjs.animate
+
+import com.greencatsoft.angularjs.InjectionTarget
+import scala.scalajs.js
+
+trait AnimateAware extends InjectionTarget {
+  implicit var animate: Animate = _
+
+  override def dependencies = super.dependencies :+ Animate.Name
+
+  override def inject(args: Seq[js.Any]) {
+    super.inject(args)
+
+    var index = dependencies.indexOf(Animate.Name) ensuring (_ >= 0)
+    this.animate = args(index).asInstanceOf[Animate]
+  }
+}
