+++ src/main/scala/com/greencatsoft/angularjs/cacheFactory/CacheFactory.scala
@@ -1,0 +1,9 @@
+package com.greencatsoft.angularjs.cacheFactory
+
+import scala.scalajs.js
+
+trait CacheFactory extends js.Function2[String,js.Object,Cache]
+
+object CacheFactory {
+  val Name = "$cacheFactory"
+}
