+++ src/main/scala/com/greencatsoft/angularjs/compile/CompileAware.scala
@@ -1,16 +1,7 @@
-package com.greencatsoft.angularjs.services
+package com.greencatsoft.angularjs.compile
 
 import com.greencatsoft.angularjs.InjectionTarget
-
 import scala.scalajs.js
-
-trait Compile extends js.Function3[js.Any,js.Function,Int,js.Function] {
-
-}
-
-object Compile {
-  val Name = "$compile"
-}
 
 trait CompileAware extends InjectionTarget {
   implicit var compile: Compile = _
