+++ src/main/scala/com/greencatsoft/angularjs/document/Document.scala
@@ -1,0 +1,5 @@
+package com.greencatsoft.angularjs.document
+
+object Document {
+ val Name = "$document"
+}
