+++ src/main/scala/com/greencatsoft/angularjs/interval/IntervalAware.scala
@@ -1,18 +1,7 @@
-package com.greencatsoft.angularjs.services
+package com.greencatsoft.angularjs.interval
 
 import com.greencatsoft.angularjs.InjectionTarget
-import com.greencatsoft.angularjs.q.Promise
-
 import scala.scalajs.js
-
-
-trait Interval extends js.Function4[js.Function,Int,Int,Boolean,Promise] {
-  def cancel(promise: Promise = null) : Boolean= ???
-}
-
-object Interval {
-  val Name = "$interval"
-}
 
 trait IntervalAware extends InjectionTarget {
   implicit var interval: Interval = _
