+++ src/main/scala/com/greencatsoft/angularjs/parse/Parse.scala
@@ -1,0 +1,9 @@
+package com.greencatsoft.angularjs.parse
+
+import scala.scalajs.js
+
+trait Parse extends js.Function1[String,js.Function]
+
+object Parse {
+  val Name = "$parse"
+}
