+++ src/main/scala/com/greencatsoft/angularjs/timeout/TimeoutAware.scala
@@ -1,17 +1,7 @@
-package com.greencatsoft.angularjs.services
+package com.greencatsoft.angularjs.timeout
 
 import com.greencatsoft.angularjs.InjectionTarget
-import com.greencatsoft.angularjs.q.Promise
-
 import scala.scalajs.js
-
-trait Timeout extends js.Function3[js.Function,Int,Boolean,Promise] {
-  def cancel(promise: Promise = null) : Boolean= ???
-}
-
-object Timeout {
-  val Name = "$timeout"
-}
 
 trait TimeoutAware extends InjectionTarget {
   implicit var timeout: Timeout = _
