+++ src/main/scala/com/greencatsoft/angularjs/window/Window.scala
@@ -1,0 +1,6 @@
+package com.greencatsoft.angularjs.window
+
+object Window {
+  val Name = "$window"
+}
+
